refactor: refactor Card props and Slice template configs
