--- packages/card/card-loading-proptypes.js
@@ -1,8 +1,0 @@
-import PropTypes from "prop-types";
-import {
-  sharedPropTypes,
-  sharedDefaultProps
-} from "./card-loading-proptypes";
-
-export const loadingPropTypes = sharedPropTypes;
-export const loadingDefaultProps = sharedDefaultProps;
